Changes full column message section
The message changes from "Pick another column" to "Pick a valid column."
This is because they might be entering values that wouldn't count as a
column. Really just a minor nit-pick.

This commit also slightly changes the last test. It adds another full
column to the list of user responses, so that I can confirm that the
"Column full!" Message repeats properly.