Fixes diagonal win bug
The diagonal win method overwrote the diagonal coordinates hashmap with
'#map!' The coordinates of a diagonal for the most recent move would
be overwritten with the board's current discs of that diagonal. This
meant that the next time the method was called for a move in an
overwritten diagonal list, the method wouldn't be able to check for a
win anymore. 'map!' has been changed to 'map,' and the discs are now
copied to a new array.